python engines (pystr / pytok) to it's own file
