+++ lib/llm/Cargo.toml
@@ -28,7 +28,6 @@ description = "Dynamo LLM Library"
 default = []
 
 cuda_kv = ["dep:cudarc", "dep:ndarray"]
-python = ["dep:pyo3-async-runtimes", "dep:pythonize"]
 sentencepiece = ["dep:sentencepiece"]
 
 [dependencies]
@@ -106,15 +105,6 @@ itertools = { version = "0.14.0" }
 minijinja = { version = "2.3.1", features = ["loader"] }
 minijinja-contrib = { version = "2.3.1", features = ["pycompat"] }
 semver = { version = "1", features = ["serde"] }
-
-# python
-pyo3-async-runtimes = { version = "0.23.0", optional = true, default-features = false, features = [
-  "attributes",
-  "testing",
-  "tokio-runtime",
-  "unstable-streams",
-] }
-pythonize = { version = "0.23", optional = true }
 
 # GGUF
 ggus = "0.4.0"
